postman: Refactor Folder into Item
Postman collection schema allows nested folders to be contained in a
collection. Previous implementation use interface{} to allow a Folder to
contain both Folder and Item, but it is hard to modify the items
underneath later.

Thus, it will be much easier if we treat a Postman folder as an Item
with .Items > 0, and a Postman item as an Item with .Items == 0.