Add release page link and configuration flag to README
As Vanadia now already have stable release and binaries for Linux and
OSX in its GitHub release page, we should explain it in README to avoid
confusion (as apparently @xinuc is not aware that I've provided the
binaries).

Custom configuration file location is also supported in v1.1.0; it was
just not documented yet, so this is the time.